[ADD]最近の達成リーチ目コレクション達成項目を追加
最大12項目描画

+++ TriSpark_Slot/_src/_header/IMenuElements.hpp
@@ -108,7 +108,7 @@ public:
 class CMenuReachCollection : public IMenuElements {
 	static const int				PAGE_MOVE = 12;
 	const int						mFontHandle;
-	size_t							mNowPage;
+	int								mNowPage;
 	const CSlotReachCollectionData& mCollectionData;
 
 public:
